Created FaceRecognizer class to handle uploads + cleanup
